Download page and Project display pages done
