--- src/componenets/AboutUs.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function AboutUs() {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default AboutUs
