--- src/componenets/Logout.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Logout() {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default Logout
