--- src/componenets/Pricing.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Pricing() {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default Pricing
