--- src/componenets/Projects.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Projects() {
-  return (
-    <div>
-      Gomma Punda
-    </div>
-  )
-}
-
-export default Projects
