+++ src/components/AboutUs.css
@@ -1,0 +1,86 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: 20px;
+    min-height: 100vh;
+    background-color: #f5f7fa;
+    font-family: 'Roboto', sans-serif;
+  }
+  
+  .image-container {
+    margin-bottom: 50px;
+    width: 300px;
+    height: 300px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-color: #fff;
+    border-radius: 15px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    overflow: hidden;
+    transition: transform 0.3s, box-shadow 0.3s;
+  }
+  
+  .image-container:hover {
+    transform: scale(1.05);
+    box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
+  }
+  
+  
+  .profile-image {
+    max-width: 100%;
+    max-height: 100%;
+    border-radius: 15px;
+  }
+  
+  
+  .download-button {
+    padding: 15px 40px;
+    margin-top: 20px;
+    font-size: 18px;
+    color: #fff;
+    background: linear-gradient(135deg, #6d5dfc 0%, #ff758c 100%);
+    border: none;
+    border-radius: 25px;
+    cursor: pointer;
+    transition: background 0.3s, transform 0.3s;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  }
+  
+  .download-button:hover {
+    background: linear-gradient(135deg, #ff758c 0%, #6d5dfc 100%);
+    transform: translateY(-2px);
+    box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
+  }
+  
+  
+  @media (max-width: 768px) {
+    .image-container {
+      width: 250px;
+      height: 250px;
+      margin-bottom: 45px; 
+    }
+  
+    .download-button {
+      font-size: 16px;
+      padding: 12px 30px;
+      margin-top: 15px; 
+    }
+  }
+  
+  @media (max-width: 480px) {
+    .image-container {
+      width: 200px;
+      height: 200px;
+      margin-bottom: 40px; 
+    }
+  
+    .download-button {
+      font-size: 14px;
+      padding: 10px 20px;
+      margin-top: 10px; 
+    }
+  }
+
