+++ src/components/Logout.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Logout() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Logout
