+++ src/components/Pricing.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Pricing() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Pricing
